\name{PanelQGARCH-Data}
\alias{Y_N_T}
\alias{Initial_tau1per_N_3}
\alias{Initial_tau2.5per_N_3}
\alias{Initial_tau97.5per_N_3}
\alias{Initial_tau99per_N_3}
\title{Data set}
\description{
  "Y_N_T" is the dataset of 96 daily log-returns of S&P100 from January 3, 2012 to June 28, 2024, are downloaded from the website of Yahoo Finance.
  "Initial_tau1per_N_3", "Initial_tau2.5per_N_3", "Initial_tau97.5per_N_3" and "Initial_tau99per_N_3" are the initial estimates in paper at the lower and upper 1\% and 2.5\% quantile levels, respectively. 
}
\author{Wenyu Li}

