\name{PanelQGARCH-package}
\alias{PanelQGARCH-package}
\alias{PanelQGARCH}
\docType{package}
\title{\packageTitle{PanelQGARCH}}
\description{\packageDescription{PanelQGARCH}}
\section{Package Content}{\packageIndices{PanelQGARCH}}
\author{\packageAuthor{PanelQGARCH}}
\section{Maintainer}{\packageMaintainer{PanelQGARCH}}
\keyword{package}
